fix(models): cast input to int32 before overlap matmul
bool @ bool in numpy saturates to True rather than summing counts,
so any column with >= 1 connected active synapse scored 1 regardless
of how many. Cast input_bool to int32 before the matmul so overlaps
are proper counts.

Also rewrote the two weak tests that accidentally passed despite the
bug: test_connected_synapses_only_count and
test_encode_stimulus_threshold_zeros_low_overlap now assert specific
winners/losers rather than just checking output size.

+++ halo/models/cortical_unit.py
@@ -105,8 +105,9 @@ class CorticalUnit(CorticalUnitBase):
             self._permanences >= self._config.syn_perm_connected
         ) & self._potential_pool  # (n_columns, input_dim) bool
 
-        # Step 2 — overlap: count of connected + active input bits per column
-        overlaps: np.ndarray = connected @ input_bool  # (n_columns,) int
+        # Step 2 — overlap: count of connected + active input bits per column.
+        # Cast to int32 so matmul sums counts, not saturated booleans.
+        overlaps: np.ndarray = connected @ input_bool.astype(np.int32)  # (n_columns,) int
 
         # Step 3 — stimulus threshold
         overlaps = np.where(
